fix dokku_app
