--- library/dokku_app.py
@@ -1,69 +1,0 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
-from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.dokku_app import (
-    dokku_app_ensure_present,
-    dokku_app_ensure_absent,
-)
-
-DOCUMENTATION = """
----
-module: dokku_app
-short_description: Create or destroy dokku apps
-options:
-  app:
-    description:
-      - The name of the app
-    required: True
-    default: null
-    aliases: []
-  state:
-    description:
-      - The state of the app
-    required: False
-    default: present
-    choices: [ "present", "absent" ]
-    aliases: []
-author: Jose Diaz-Gonzalez
-requirements: [ ]
-"""
-
-EXAMPLES = """
-- name: Create a dokku app
-  dokku_app:
-    app: hello-world
-
-- name: Delete that repo
-  dokku_app:
-    app: hello-world
-    state: absent
-"""
-
-
-def main():
-    fields = {
-        "app": {"required": True, "type": "str"},
-        "state": {
-            "required": False,
-            "default": "present",
-            "choices": ["present", "absent"],
-            "type": "str",
-        },
-    }
-    choice_map = {
-        "present": dokku_app_ensure_present,
-        "absent": dokku_app_ensure_absent,
-    }
-
-    module = AnsibleModule(argument_spec=fields, supports_check_mode=False)
-    is_error, has_changed, result = choice_map.get(module.params["state"])(
-        module.params
-    )
-
-    if is_error:
-        module.fail_json(msg=result["error"], meta=result)
-    module.exit_json(changed=has_changed, meta=result)
-
-
-if __name__ == "__main__":
-    main()
